FlashAttention needs contiguous gradients (#548)
* FlashAttention needs contiguous gradients

This was not visible before because the test was creating tensors of the wrong dtype, which when casted to the right dtype would yield contiguous tensors

* Fix lint

+++ tests/test_mem_eff_attention.py
@@ -605,7 +605,9 @@ def test_backward(
 
     grad_out = torch.ones_like(out)
     if grad_out_contiguous is False:
-        grad_out = torch.tensor([1.0], device=device)[None, None, :].expand_as(out)
+        grad_out = torch.tensor([1.0], dtype=query.dtype, device=device)[
+            None, None, :
+        ].expand_as(out)
 
     out.backward(grad_out)
     del out
